feat: Add certificate and GitHub secret management, refactor signing by removing Fulcio/Sigstore integration, and update project version.

+++ hooks/post-build/9900.CLEAN_UP.sh
@@ -2,7 +2,12 @@
 
 . "$KAM_HOOKS_ROOT/lib/utils.sh"
 
-# log_warn " comment out to enable !" && exit 0
+log_warn " comment out to enable !" && exit 0
+
+if [ "$KAM_RELEASE_ENABLED" != "1" ]; then
+    log_info "KAM_RELEASE_ENABLED != 1, skipping clean up"
+    exit 0
+fi
 
 rm -r "$KAM_DIST_DIR" || log_error "failed to clean up."
